renamed CartItem.js to CartItem_temp.js
